ログイン状態での場合分け（修正）

+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :set_item, only: [:show, :edit, :update]
+  before_action :set_item, only: [:show, :edit, :update, :destroy]
 
   def index
     @items = Item.all.order(id: "DESC")
@@ -38,6 +38,9 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    unless @item.user_id == current_user.id
+      redirect_to action: :index
+    end
     item = Item.find(params[:id])
     item.destroy
     redirect_to root_path
